Project Switch to LWRP & 2D Renderer
Huge update that affects every prefab in the game due to the materials that changed. If there is a conflict within prefabs which you have important changes, you can use yours and then change the materials to the new material called "Sprite-Lit Default". 

Other things that has changes are the project settings. These are super important otherwise you will not be able to see anything in the game. 

Finally, the Agata Tag was for some reason missing in the Tag Manager. So I have added that also.